+++ docs/web/api/doc/3.0.0/text/transformer-register.tpl.js
@@ -1,0 +1,12 @@
+
+// register a transformer function
+_.xng.transform("json_string", function(str) {
+	return JSON.parse(str);
+});
+
+// register multiple transformer functions
+_.xng.transform({
+	"json_string" : function(str) {
+		return JSON.parse(str);
+	}
+});